# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Crousel Ecommerce API
  description: API autentikasi customer/admin, pengaturan toko, katalog, dan notifikasi untuk aplikasi ecommerce.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /admin/customization:
    put:
      tags: [Settings]
      summary: Simpan kustomisasi tampilan toko
      description: Menyimpan warna secondary untuk diterapkan di frontend. Harus format hex 6 digit (contoh #ff5500).
      operationId: updateCustomization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secondary_color]
              properties:
                secondary_color: { type: string, pattern: '^#[0-9a-fA-F]{6}$', example: '#ff5500' }
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CustomizationUploadRequest'
      responses:
        '200':
          description: Kustomisasi berhasil disimpan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Kustomisasi berhasil disimpan. }
                  data:
                    type: object
                    properties:
                      secondary_color: { type: string, example: '#ff5500' }
                      hero_image: { type: string, format: uri, nullable: true }
                      size_guide_image: { type: string, format: uri, nullable: true }
                      about_image: { type: string, format: uri, nullable: true }
                      story_images: { type: array, minItems: 5, maxItems: 5, items: { type: string, format: uri, nullable: true } }
        '422': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [Settings]
      summary: Simpan kustomisasi beserta gambar
      description: Upload gambar hero, panduan ukuran, tentang kami, dan tepat lima gambar cerita menggunakan multipart/form-data.
      operationId: uploadCustomizationImages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/CustomizationUploadRequest' }
      responses:
        '200': { description: Kustomisasi berhasil disimpan }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Gunakan token dari endpoint register/login sebagai `Bearer <token>`.
...
  schemas:
...
    CustomizationUploadRequest:
      type: object
      properties:
        secondary_color: { type: string, pattern: '^#[0-9a-fA-F]{6}$', example: '#ff5500' }
        hero_image: { type: string, format: binary, description: 'Gambar hero maksimal 5 MB' }
        size_guide_image: { type: string, format: binary, description: 'Gambar panduan ukuran maksimal 5 MB' }
        about_image: { type: string, format: binary, description: 'Gambar tentang kami maksimal 5 MB' }
        story_images:
          type: array
          minItems: 5
          maxItems: 5
          items: { type: string, format: binary, description: 'Gambar cerita maksimal 5 MB per file' }
...
    ValidationError:
      type: object
      properties:
        message: { type: string, example: The given data was invalid. }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example:
            whatsapp: [The whatsapp has already been taken.]
    UnauthenticatedError:
      type: object
      properties:
        message: { type: string, example: Unauthenticated. }
  responses:
    ValidationError:
      description: Data tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
    Unauthenticated:
      description: Token tidak ada atau tidak valid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UnauthenticatedError' }
...